+++ src/payloads/leave_chat.rs
@@ -8,12 +8,12 @@
 // [`schema`]: https://github.com/WaffleLapkin/tg-methods-schema
 use serde::Serialize;
 
-use crate::types::ChatId;
+use crate::types::{ChatId, True};
 
 impl_payload! {
     /// Use this method for your bot to leave a group, supergroup or channel. Returns _True_ on success.
     #[derive(Debug, PartialEq, Eq, Hash, Clone, Serialize)]
-    pub LeaveChat (LeaveChatSetters) => String {
+    pub LeaveChat (LeaveChatSetters) => True {
         required {
             /// Unique identifier for the target chat or username of the target channel (in the format `@channelusername`)
             pub chat_id: ChatId [into],