+++ src/payloads/set_chat_title.rs
@@ -8,12 +8,12 @@
 // [`schema`]: https://github.com/WaffleLapkin/tg-methods-schema
 use serde::Serialize;
 
-use crate::types::ChatId;
+use crate::types::{ChatId, True};
 
 impl_payload! {
     /// Use this method to change the title of a chat. Titles can't be changed for private chats. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Returns _True_ on success.
     #[derive(Debug, PartialEq, Eq, Hash, Clone, Serialize)]
-    pub SetChatTitle (SetChatTitleSetters) => String {
+    pub SetChatTitle (SetChatTitleSetters) => True {
         required {
             /// Unique identifier for the target chat or username of the target channel (in the format `@channelusername`)
             pub chat_id: ChatId [into],